requirements.txt
google-generativeai
pillow
pypdfium2
markdown
